requirements.txt
django-push-notifications
aioredis==1.3.1
amqp==2.6.0
apns2==0.7.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==20.2.0
billiard==3.6.3.0
boto==2.49.0
boto3==1.10.46
botocore==1.13.46
celery==4.4.4
certifi==2019.11.28
cffi==1.14.0
cfscrape==2.1.1
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.1
django-ajax-selects==1.9.1
django-allauth==0.42.0
django-appconf==1.0.3
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-debug-toolbar==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-push-notifications==2.0.0
django-rest-auth==0.9.5
django-storages==1.8
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.2.0
docutils==0.15.2
drf-yasg==1.17.1
fcm-django==0.3.4
h2==2.6.2
hiredis==1.0.1
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==2.8
imgkit==1.0.2
importlib-metadata==1.6.0
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.10
MarkupSafe==1.1.1
msgpack==0.6.2
munch==2.5.0
mypy==0.782
mypy-extensions==0.4.3
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.1
pilkit==2.0
Pillow==6.2.1
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyfcm==1.4.7
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.11.0
raven==6.10.0
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
xlrd==1.2.0
zipp==3.1.0
zope.interface==4.7.1
